% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_creation_functions.R
\name{binary_logit}
\alias{binary_logit}
\title{Binary variable from a binomial distribution with a logit link}
\usage{
binary_logit(data)
}
\arguments{
\item{data}{vector, scalar, matrix, or data.frame representing the latent variable used to draw the binary outcome}
}
\description{
Binary variable from a binomial distribution with a logit link
}
