% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_creation_functions.R
\name{binomial_count}
\alias{binomial_count}
\title{Count variable from a binomial distribution with a logit link}
\usage{
binomial_count(data, k)
}
\arguments{
\item{data}{vector, scalar, matrix, or data.frame representing the latent variable used to draw the count outcome}

\item{k}{number of binomial trials, i.e. maximum of the count variable}
}
\description{
Count variable from a binomial distribution with a logit link
}
\examples{

fabricate_data(N = 10, Y1 = rnorm(N),  Y2 = binomial_count(Y1, 4))
fabricate_data(N = 10, Y1 = rnorm(N),  Y2 = binomial_count(Y1, 2))

binomial_count(runif(50), 3)
}
