% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabricate.R
\name{cross_level}
\alias{cross_level}
\title{Creates cross-classified (partially non-nested, joined data) with a fixed
correlation structure.}
\usage{
cross_level(N = NULL, ID_label = NULL, working_environment_ = NULL,
  by = NULL, ..., data_arguments = quos(...))
}
\arguments{
\item{N}{(required) The number of observations in the resulting data frame}

\item{ID_label}{Internal keyword used to sepcify the name of the ID variable created for the new level. If left empty, this will be the name the level is assigned to as part of a \code{fabricate()} call.}

\item{working_environment_}{Internal keyword not for end user use.}

\item{by}{The result of a call to \code{join()} which specifies how the cross-classified data will be created}

\item{...}{A variable or series of variables to add to the resulting data frame after the cross-classified data is created.}

\item{data_arguments}{Internal keyword not for end user use.}
}
\description{
Creates cross-classified (partially non-nested, joined data) with a fixed
correlation structure.
}
