% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_variable.R
\name{describe_variable}
\alias{describe_variable}
\title{Describe Variable}
\usage{
describe_variable(x)
}
\arguments{
\item{x}{a vector representing a single variable}
}
\value{
a data.frame summary of the variable
}
\description{
Describe Variable
}
\examples{

var1 <- sample(1:5, 50, replace = TRUE)
describe_variable(var1)

var2 <- runif(50)
describe_variable(var2)
var3 <- rep(letters[1:5], each = 5)
describe_variable(var3)

var4 <- rep(c(TRUE, FALSE), c(70, 30))
describe_variable(var4)


}
