% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabricate_data.R
\name{fabricate_data}
\alias{fabricate_data}
\title{Fabricate data}
\usage{
fabricate_data(data = NULL, N = NULL, ID_label = NULL, ...)
}
\arguments{
\item{data}{(optional) user-provided data that forms the basis of the fabrication, i.e. you can add variables to existing data. Provide either \code{N} or \code{data} (\code{N} is the number of rows of the data if \code{data} is provided).}

\item{N}{(optional) number of units to draw, if drawing a single level of data (i.e. not hierarchical data)}

\item{ID_label}{(optional) variable name for ID variable, i.e. citizen_ID}

\item{...}{Variable or level-generating arguments, such as \code{my_var = rnorm(N)}. You may also pass \code{level()} arguments, which define a level of a multi-level dataset. For example, you could send to \code{...} \code{my_level = level(N = 5, var = rnorm)}. See examples.}
}
\value{
data.frame
}
\description{
\code{fabricate_data} helps you simulate a dataset before you collect it. You can either start with your own data and add simulated variables to it (by passing \code{data} to \code{fabricate_data()}) or start from scratch by defining \code{N}. Create hierarchical data with multiple levels of data such as citizens within cities within states using \code{level()}. You can use any R function to create each variable. We provide several built-in options to easily draw from binary and count outcomes, \code{\link{draw_binary}} and \code{\link{draw_count}}.
}
\examples{

# Draw a single-level dataset with no covariates
df <- fabricate_data(N = 100)
head(df)

# Draw a single-level dataset with a covariate
df <- fabricate_data(
  N = 100,
  height_ft = runif(N, 3.5, 8)
)
head(df)

# Start with existing data
df <- fabricate_data(
  data = df,
  new_variable = rnorm(N)
)

# Draw a two-level hierarchical dataset
# containing cities within regions
df <- fabricate_data(
 regions = level(N = 5),
 cities = level(N = 2, pollution = rnorm(N, mean = 5)))
head(df)

# Start with existing data and add variables to hierarchical data
# note: do not provide N when adding variables to an existing level
df <- fabricate_data(
  data = df,
  regions = level(watershed = sample(c(0, 1), N, replace = TRUE)),
  cities = level(runoff = rnorm(N))
)

}
