% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabricate_rewrite.R
\name{fabricate_revised}
\alias{fabricate_revised}
\title{Fabricate data}
\usage{
fabricate_revised(data = NULL, N = NULL, ID_label = NULL, ...)
}
\arguments{
\item{data}{(optional) user-provided data that forms the basis of the fabrication, i.e. you can add variables to existing data. Provide either \code{N} or \code{data} (\code{N} is the number of rows of the data if \code{data} is provided).}

\item{N}{(optional) number of units to draw. If provided as \code{fabricate(N = 5)}, this determines the number of units in the single-level data. If provided in \code{level}, i.e. \code{fabricate(cities = level(N = 5))}, \code{N} determines the number of units in a specific level of a hierarchical dataset.}

\item{ID_label}{(optional) variable name for ID variable, i.e. citizen_ID}

\item{...}{Variable or level-generating arguments, such as \code{my_var = rnorm(N)}. For \code{fabricate}, you may also pass \code{level()} arguments, which define a level of a multi-level dataset. See examples.}
}
\value{
data.frame
}
\description{
\code{fabricate_revised} helps you simulate a dataset before you collect it. You can either start with your own data and add simulated variables to it (by passing \code{data} to \code{fabricate()}) or start from scratch by defining \code{N}. Create hierarchical data with multiple levels of data such as citizens within cities within states using \code{level()}. You can use any R function to create each variable. We provide several built-in options to easily draw from binary and count outcomes, \code{\link{draw_binary}} and \code{\link{draw_discrete}}.
}
\examples{

# Draw a single-level dataset with no covariates
df <- fabricate_revised(N = 100)
head(df)

# Draw a single-level dataset with a covariate
df <- fabricate_revised(
  N = 100,
  height_ft = runif(N, 3.5, 8)
)
head(df)

# Start with existing data
df <- fabricate_revised(
  data = df,
  new_variable = rnorm(N)
)

# Draw a two-level hierarchical dataset
# containing cities within regions
df <- fabricate_revised(
 regions = add_level_new(N = 5),
 cities = nest_level_new(N = 2, pollution = rnorm(N, mean = 5)))
head(df)

# Start with existing data and add variables to hierarchical data
# note: do not provide N when adding variables to an existing level
df <- fabricate_revised(
  data = df,
  regions = modify_level_new(watershed = sample(c(0, 1), N, replace = TRUE)),
  cities = modify_level_new(runoff = rnorm(N))
)

}
