% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabricatr.R
\docType{package}
\name{fabricatr}
\alias{fabricatr}
\alias{fabricatr-package}
\title{fabricatr package}
\description{
fabricatr helps you imagine your data before you collect it. Hierarchical
data structures and correlated data can be easily simulated, either from
random number generators or by resampling from existing data sources.
}
