% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_unique_variables_by_level}
\alias{get_unique_variables_by_level}
\title{Find which variables are unique at a given level in hierarchical data}
\usage{
get_unique_variables_by_level(data, ID_label, superset = NULL)
}
\arguments{
\item{data}{a data.frame}

\item{ID_label}{the ID label to split upon}

\item{superset}{Superset contains a vector of character strings that contain variables
the modify level call is going to write. Some of these may be columns
in the data frame, others might not be. If superset is specified,
then we definitely only want to check those variables}
}
\value{
a character vector enumerating the unique variables
}
\description{
Find which variables are unique at a given level in hierarchical data
}
