% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_creation_functions.R
\name{logit}
\alias{logit}
\title{Simple helper function wrapping the plogis function.}
\usage{
logit(x)
}
\arguments{
\item{x}{Latent variable to be logit-transformed}
}
\description{
Simple helper function wrapping the plogis function.
}
\keyword{internal}
