% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_creation_functions.R
\name{probit}
\alias{probit}
\title{Simple helper function wrapping the pnorm function.}
\usage{
probit(x)
}
\arguments{
\item{x}{Latent variable to be probit-transformed}
}
\description{
Simple helper function wrapping the pnorm function.
}
\keyword{internal}
