% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_data.R
\docType{data}
\name{ALL}
\alias{ALL}
\title{Magic constant: pass all units at a resampling level}
\format{
An object of class \code{integer} of length 1.
}
\usage{
ALL
}
\description{
Use \code{ALL} in the \code{N} argument of \code{resample_data} to keep
every unit at a given hierarchical level while still resampling inner
levels.
}
\keyword{datasets}
