% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_data.R
\docType{data}
\name{ALL}
\alias{ALL}
\title{Magic number constant to allow users to specify \code{ALL} for passthrough
resampling}
\format{
An object of class \code{integer} of length 1.
}
\usage{
ALL
}
\description{
Magic number constant to allow users to specify \code{ALL} for passthrough
resampling
}
\keyword{internal}
