% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_creation_functions.R
\name{binomial_count}
\alias{binomial_count}
\title{Count variable from a binomial distribution with a logit link}
\usage{
binomial_count(data, k)
}
\arguments{
\item{data}{}

\item{k}{number of binomial trials, i.e. maximum of the count variable}
}
\description{
Count variable from a binomial distribution with a logit link
}
