% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_creation_functions.R
\name{correlate}
\alias{correlate}
\title{Perform generation of a correlated random variable.}
\usage{
correlate(draw_handler, ..., given, rho)
}
\arguments{
\item{draw_handler}{The unquoted name of a function to generate data.
Currently, `draw_binary`, `draw_binomial`, and `draw_count` are supported.}

\item{...}{The arguments to draw_handler (e.g. `prob`, `mean`, etc.)}

\item{given}{A vector that can be ordered; the reference distribution X that
Y will be correlated with.}

\item{rho}{A rank correlation coefficient between -1 and 1.}
}
\description{
In order to generate a random variable of a specific distribution based on
another variable of any distribution and a correlation coefficient `rho`,
we map the first, known variable into the standard normal space via affine
transformation, and generate a conditional distribution of Y. This function
}
