% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_level.R
\name{cross_level}
\alias{cross_level}
\title{Creates cross-classified (partially non-nested, joined data) with a fixed
correlation structure.}
\usage{
cross_level(N = NULL, by = NULL, ...)
}
\arguments{
\item{N}{(required) The number of observations in the resulting data frame}

\item{by}{The result of a call to \code{join()} which specifies how the
cross-classified data will be created}

\item{...}{A variable or series of variables to add to the resulting data
frame after the cross-classified data is created.}
}
\value{
data.frame
}
\description{
Creates cross-classified (partially non-nested, joined data) with a fixed
correlation structure.
}
\examples{

# Generate cross-classified data and merge, no correlation
students <- fabricate(
 primary_school = add_level(N = 20, ps_quality = runif(N, 1, 10)),
 secondary_school = add_level(N = 15, ss_quality = runif(N, 1, 10), nest=FALSE),
 students = cross_level(N = 500, by = join(primary_school, secondary_school))
)
head(students)

# Cross-classified data with a correlation structure
students <- fabricate(
 primary_school = add_level(N = 20, ps_quality = runif(N, 1, 10)),
 secondary_school = add_level(N = 15, ss_quality = runif(N, 1, 10), nest=FALSE),
 students = cross_level(N = 500, by = join(ps_quality, ss_quality, rho = 0.5))
)
cor(students$ps_quality, students$ss_quality)

}
