% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_creation_functions.R
\name{draw_binary}
\alias{draw_binary}
\title{Binary variable from a binomial distribution}
\usage{
draw_binary(vector, link = "logit")
}
\arguments{
\item{vector}{vector representing either the latent variable used to draw the count outcome (if link is "logit" or "probit") or the probability for the count outcome (if link is "identity")}

\item{link}{link function between the latent variable and the probability of a postiive outcome, i.e. "logit", "probit", or "identity". For the "identity" link, the latent variable must be a probability.}
}
\description{
Binary variable from a binomial distribution
}
\examples{

fabricate_data(N = 10, Y1 = rnorm(N),  Y2 = draw_binary(Y1))
fabricate_data(N = 10, Y1 = rnorm(N),  Y2 = draw_binary(Y1, link = "logit"))
fabricate_data(N = 10, Y1 = rnorm(N),  Y2 = draw_binary(Y1, link = "probit"))
fabricate_data(N = 10, Y1 = runif(N, 0, 1),  Y2 = draw_binary(Y1, link = "identity"))

draw_binary(rnorm(10))
draw_binary(rnorm(10), link = "logit")
draw_binary(rnorm(10), link = "probit")
draw_binary(runif(10, 0, 1), link = "identity")

}
