% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_binary_icc.R
\name{draw_binary_icc}
\alias{draw_binary_icc}
\title{Draw binary data with fixed intra-cluster correlation.}
\usage{
draw_binary_icc(x = 0.5, N = NULL, cluster_ids, rho = 0.5)
}
\arguments{
\item{x}{A number or vector of numbers, one probability per cluster.}

\item{N}{(Optional) A number indicating the number of observations to be
generated. Must be equal to length(cluster_ids) if provided.}

\item{cluster_ids}{A vector of factors or items that can be coerced to
clusters; the length will determine the length of the generated data.}

\item{rho}{A number indicating the desired RCC.}
}
\value{
A vector of binary numbers corresponding to the observations from
the supplied cluster IDs.
}
\description{
Data is generated to ensure inter-cluster correlation 0, intra-cluster
correlation in expectation \eqn{\rho}. Algorithm taken from Hossein,
Akhtar. "ICCbin: An R Package Facilitating Clustered Binary Data
Generation, and Estimation of Intracluster Correlation Coefficient (ICC)
for Binary Data".
}
\examples{
cluster_ids = rep(1:5, 10)
draw_binary_icc(cluster_ids = cluster_ids)
draw_binary_icc(x = 0.5, cluster_ids = cluster_ids, rho = 0.5)

}
