% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_binary_icc.R
\name{draw_binary_icc}
\alias{draw_binary_icc}
\title{Draw binary data with fixed intra-cluster correlation.}
\usage{
draw_binary_icc(x = 0.5, N = NULL, cluster_ids, rho = 0.5)
}
\arguments{
\item{x}{A number or vector of numbers, one probability per cluster.}

\item{N}{(Optional) A number indicating the number of observations to be
generated. Must be equal to length(cluster_ids) if provided.}

\item{cluster_ids}{A vector of factors or items that can be coerced to
clusters; the length will determine the length of the generated data.}

\item{rho}{A number indicating the desired RCC.}
}
\value{
A vector of binary numbers corresponding to the observations from
the supplied cluster IDs.
}
\description{
Data is generated according to the following algorithm, where \eqn{i} is
the index of a cluster and \eqn{j} is the index of a unit: \deqn{z_i ~
Bernoulli(p_i) \cr
y_{ij} ~ Bernoulli(p_{ij}) \cr
u_{ij} ~ Bernoulli(sqrt(\rho)) \cr
x_{ij} = (u_{ij}) z_i + (1 - u_{ij}) y_{ij}}
}
\details{
This system of equations ensures inter-cluster correlation 0, intra-cluster
correlation in expectation \eqn{\rho}. Algorithm from Hossein, Akhtar.
"ICCbin: An R Package Facilitating Clustered Binary Data Generation, and
Estimation of Intracluster Correlation Coefficient (ICC) for Binary Data".
We rederived the analytical properties of this data and did a simulation
study to confirm that the data generated ensured the ICC we mentioned.
}
\examples{
cluster_ids = rep(1:5, 10)
draw_binary_icc(cluster_ids = cluster_ids)
draw_binary_icc(x = 0.5, cluster_ids = cluster_ids, rho = 0.5)

}
