% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_creation_functions.R
\name{draw_binomial}
\alias{draw_binomial}
\alias{draw_categorical}
\alias{draw_ordered}
\alias{draw_count}
\alias{draw_binary}
\title{Draw discrete variables including binary, binomial count, poisson count,
ordered, and categorical}
\usage{
draw_binomial(prob, trials = 1, N = length(prob), link = "identity")

draw_categorical(prob, N = NULL, link = "identity")

draw_ordered(x, breaks = c(-1, 0, 1), break_labels = NULL, N = length(x),
  link = "identity")

draw_count(mean, N = length(mean), link = "identity")

draw_binary(prob, N = length(prob), link = "identity")
}
\arguments{
\item{prob}{A number or vector of numbers representing the probability for
binary or binomial outcomes; or a number, vector, or matrix of numbers
representing probabilities for categorical outcomes. If you supply a link
function, these underlying probabilities will be transformed.}

\item{trials}{for `draw_binomial`, the number of trials for each observation}

\item{N}{number of units to draw. Defaults to the length of the vector of
probabilities or latent data you provided}

\item{link}{link function between the latent variable and the probability of
a postiive outcome, e.g. "logit", "probit", or "identity". For the "identity"
link, the latent variable must be a probability.}

\item{x}{for `draw_ordered`, the latent data for each observation.}

\item{breaks}{vector of breaks to cut a latent outcome into ordered categories}

\item{break_labels}{vector of labels for the breaks to cut a latent outcome
into ordered categories.}

\item{mean}{for `draw_count`, the mean number of count units for each observation}
}
\description{
Drawing discrete data based on probabilities or latent traits is a common
task that can be cumbersome. Each function in our discrete drawing set creates
a different type of discrete data: `draw_binary` creates binary 0/1 data,
`draw_binomial` creates binomial data (repeated trial binary data),
`draw_categorical` creates categorical data, `draw_ordered` transforms latent
data into observed ordered categories, `draw_count` creates count data
(poisson-distributed)
}
\examples{
fabricate(N = 3,
   p = c(0, .5, 1),
   binary = draw_binary(prob = p))


fabricate(N = 3,
   x = 10 * rnorm(N),
   binary = draw_binary(prob = x, link = "probit"))

fabricate(N = 3,
   p = c(0, .5, 1),
   binomial = draw_binomial(prob = p, trials = 10))

fabricate(N = 3,
   x = 5 * rnorm(N),
   ordered = draw_ordered(x=x,
                          breaks = c(-Inf, -1, 1, Inf)))

fabricate(N = 3,
   x = c(0,5,100),
   count = draw_count(mean=x))

# Categorical
fabricate(N = 6, p1 = runif(N), p2 = runif(N), p3 = runif(N),
          cat = draw_categorical(cbind(p1, p2, p3)))

}
