% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_creation_functions.R
\name{draw_count}
\alias{draw_count}
\title{Draw a count variable from a binomial distribution}
\usage{
draw_count(latent = NULL, prob = NULL, k, link = "logit")
}
\arguments{
\item{latent}{vector representing the latent variable used to draw the count outcome (if link is "logit" or "probit")}

\item{prob}{vector representing the probability for the count outcome (if link is "identity"). Link is automatically set to "identity" if probabilities is provided.}

\item{k}{number of binomial trials, i.e. maximum of the count variable}

\item{link}{link function between the latent variable and the probability of a postiive outcome, i.e. "logit", "probit", or "identity". For the "identity" link, the latent variable must be a probability.}
}
\description{
Drawing count variables based on probabilities or latent traits is a common task that can be cumbersome.
}
\examples{

fabricate(N = 10, Y1 = rnorm(N),  Y2 = draw_count(Y1, k = 4))
fabricate(N = 10, Y1 = rnorm(N),  Y2 = draw_count(Y1, k = 4, link = "probit"))
fabricate(N = 10, Y1 = runif(N, 0, 1),  Y2 = draw_count(prob = Y1, k = 4))

draw_count(rnorm(10), k = 4)
draw_count(rnorm(10), k = 4, link = "logit")
draw_count(rnorm(10), k = 4, link = "probit")
draw_count(prob = runif(10, 0, 1), k = 4)
draw_count(prob = runif(10, 0, 1), k = 4, link = "identity")
}
