% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_multivariate.R
\name{draw_multivariate}
\alias{draw_multivariate}
\title{Draw multivariate random variables}
\usage{
draw_multivariate(formula, sep = "_")
}
\arguments{
\item{formula}{Formula describing the multivariate draw. The lefthand side is the names or prefix and the right-hand side is the multivariate draw function call, such as mvrnorm from the MASS library or rmnom from the extraDistr library.}

\item{sep}{Separator string between prefix and variable number. Only used when a single character string is provided and multiple variables created.}
}
\value{
tibble
}
\description{
Draw multivariate random variables
}
\examples{

library(extraDistr)
# draw from multinomial (multivariate binomial) distribution
draw_multivariate(Y ~ rmnom(5, size = 3, prob = c(0.25, 0.1)))
draw_multivariate(c(Y_1, Y_2) ~ rmnom(5, size = 3, prob = c(0.25, 0.1)))

}
