% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_normal_icc.R
\name{draw_normal_icc}
\alias{draw_normal_icc}
\title{Draw normal data with fixed intra-cluster correlation.}
\usage{
draw_normal_icc(x = 0, N = NULL, cluster_ids, sd = 1, rho = 0.5)
}
\arguments{
\item{x}{A number or vector of numbers, one mean per cluster.}

\item{N}{(Optional) A number indicating the number of observations to be
generated. Must be equal to length(cluster_ids) if provided.}

\item{cluster_ids}{A vector of factors or items that can be coerced to
clusters; the length will determine the length of the generated data.}

\item{sd}{A number or vector of numbers, indicating the standard deviation of
each cluster's error terms}

\item{rho}{A number indicating the desired RCC.}
}
\value{
A vector of numbers corresponding to the observations from
the supplied cluster IDs.
}
\description{
Data is generated according to the following algorithm, where \eqn{i} is
the index of a cluster and \eqn{j} is the index of a unit:
}
\details{
\deqn{\sigma^2_{\alpha}(i) = (\rho \sigma^2_{\epsilon}(i)) / (1 - \rho) \cr
\mu_{ij} ~ \mathcall{N}(\mu_i, \sigma_{\epsilon}(i))
\alpha_{i} ~ \mathcal{N}(0, \sigma_{\alpha}(i)) \cr
x_{ij} = \mu_{ij} + \alpha_{i}}

This system of equations ensures inter-cluster correlation 0, intra-cluster
correlation in expectation \eqn{\rho}. Algorithm discussed at
\url{https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc}
}
\examples{
cluster_ids = rep(1:5, 10)
draw_normal_icc(cluster_ids = cluster_ids)

}
