% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_normal_icc.R
\name{draw_normal_icc}
\alias{draw_normal_icc}
\title{Draw normal data with fixed intra-cluster correlation.}
\usage{
draw_normal_icc(x = 0, N = NULL, clusters, sd = 1, rho = 0.5)
}
\arguments{
\item{x}{A number or vector of numbers, one mean per cluster.}

\item{N}{(Optional) A number indicating the number of observations to be
generated. Must be equal to length(clusters) if provided.}

\item{clusters}{A vector of factors or items that can be coerced to
clusters; the length will determine the length of the generated data.}

\item{sd}{A number or vector of numbers, indicating the standard deviation of
each cluster's error terms}

\item{rho}{A number indicating the desired RCC.}
}
\value{
A vector of numbers corresponding to the observations from
the supplied cluster IDs.
}
\description{
Data is generated to ensure inter-cluster correlation 0, intra-cluster
correlation in expectation \eqn{\rho}{rho}. The data generating process
used in this function is specified at the following URL:
\url{https://stats.stackexchange.com/questions/263451/create-synthetic-data-with-a-given-intraclass-correlation-coefficient-icc}
}
\examples{
clusters = rep(1:5, 10)
draw_normal_icc(clusters = clusters)

}
