% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_categorical.R
\name{draw_quantile}
\alias{draw_quantile}
\title{Draw quantile bucket assignments}
\usage{
draw_quantile(type, N)
}
\arguments{
\item{type}{Number of buckets.}

\item{N}{Number of observations.}
}
\value{
Ordered factor.
}
\description{
Draws N observations and assigns them to \code{type} equally sized
quantile buckets via a uniform latent variable.
}
\examples{
draw_quantile(type = 4, N = 100)

}
