% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabricate_data.R
\name{fabricate_data}
\alias{fabricate_data}
\title{Fabricate data f}
\usage{
fabricate_data(..., N = NULL, ID_label = NULL, data = NULL)
}
\arguments{
\item{...}{Data generating arguments, such as \code{my_var = rnorm(N)}. You may also pass \code{level()} arguments, which define a level of a multi-level dataset. For example, you could send to \code{...} \code{level(my_level, var = rnorm)}. See examples.}

\item{N}{number of units to draw}

\item{ID_label}{label for ID variable, i.e. citizen_ID}

\item{data}{user-provided data that forms the basis of the fabrication, i.e. you can add variables to existing data. Provide either \code{N} or \code{data} (\code{N} is the number of rows of the data if \code{data} is provided).}
}
\value{
data.frame
}
\description{
fff
}
\examples{

}
