% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level.R
\name{level}
\alias{level}
\title{Fabricate a Level of Data for Hierarchical Data}
\usage{
level(N = NULL, ...)
}
\arguments{
\item{N}{number of units to draw in the level}

\item{...}{Data generating arguments, such as \code{my_var = rnorm(N)}. You may also pass \code{level()} arguments, which define a level of a multi-level dataset. For example, you could send to \code{...} \code{level(my_level, var = rnorm)}. See examples.}
}
\description{
Fabricate a Level of Data for Hierarchical Data
}
\examples{

# Draw a two-level hierarchical dataset
# containing cities within regions
df <- fabricate(
 regions = level(N = 5),
 cities = level(N = 10, pollution = rnorm(N, mean = 5)))
head(df)

}
