% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level.R
\name{level}
\alias{level}
\title{Fabricate a Level of Data for Multi-Level Hierarchical Data}
\usage{
level(ID_label, N = NULL, ..., by = NULL, data = NULL)
}
\arguments{
\item{ID_label}{variable name for ID variable, i.e. citizen_ID (optional)}

\item{N}{number of units to draw in the level}

\item{...}{Data generating arguments, such as \code{my_var = rnorm(N)}. You may also pass \code{level()} arguments, which define a level of a multi-level dataset. For example, you could send to \code{...} \code{level(my_level, var = rnorm)}. See examples.}

\item{by}{name of variable to merge by to level above, if the level is not at the top of the hierarchy.}

\item{data}{user-provided data that forms the basis of the fabrication at this level, i.e. you can add variables to existing data. Provide either \code{N} or \code{data} (\code{N} is the number of rows of the data if \code{data} is provided).}
}
\description{
Fabricate a Level of Data for Multi-Level Hierarchical Data
}
\examples{

# Draw a two-level hierarchical dataset
# containing cities within regions
df <- fabricate_data(
 regions = level(N = 5),
 cities = level(N = 10, pollution = rnorm(N, mean = 5)))
head(df)

# Use existing data in a level using data
region_data <- data.frame(capital = c(1, 0, 0, 0, 0))
fabricate_data(regions = level(data = region_data,
                               gdp = runif(5)),
               cities = level(N = 5,
                              subways = rnorm(N, mean = 5)))

}
