% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_data.R
\name{resample_data}
\alias{resample_data}
\title{Resample data, including multi-level data}
\usage{
resample_data(data, N, ID_labels = NULL)
}
\arguments{
\item{data}{A data.frame, usually provided by the user.}

\item{N}{The number of sample N to return. If multiple levels are defined in ID_labels, N should be a vector.}

\item{ID_labels}{The variables that indicate the data hierarchy.}
}
\value{
A data.frame
}
\description{
Resample data, including multi-level data
}
