% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_data.R
\name{resample_data}
\alias{resample_data}
\title{Resample data, including hierarchical data}
\usage{
resample_data(data, N, ID_labels = NULL)
}
\arguments{
\item{data}{A data.frame, usually provided by the user.}

\item{N}{The number of sample N to return. If N is a single scalar and no labels are provided, N will specify the number of unit observations to resample. If N is named, or if the ID_labels argument is specified (in which case, both N and ID_labels should be the same length), then the units resampled will be values of the levels resampled (this is useful for, e.g., cluster resampling). If N is the constant ALL for any level, all units of this level will be transparently passed through to the next level of resampling.}

\item{ID_labels}{A character vector of the variables that indicate the data hierarchy, from highest to lowest (i.e., from cities to citizens).}
}
\value{
A data.frame
}
\description{
Resample data, including hierarchical data
}
\examples{

# Bootstrap a dataset without any hierarchy. N specifies a number of observations to return

baseline_survey <- fabricate(N = 5, Y_pre = rnorm(N))
bootsrapped_data <- resample_data(baseline_survey, N = 10)
bootsrapped_data

# Resample by a single level of a hierarchical dataset (e.g. resampling clusters of observations)
# N specifies a number of clusters to return

clustered_survey <- fabricate(
  clusters = level(N=25),
  cities = level(N=runif(25, 1, 5), population=runif(n = N, min=50000, max=1000000))
)

# Specify the name of the cluster variable one of two ways

cluster_resample <- resample_data(clustered_survey, N = 5, ID_labels = "clusters")
cluster_resample

cluster_resample_2 <- resample_data(clustered_survey, N=c(clusters = 5))
cluster_resample_2

# Resample a hierarchical dataset on multiple levels

my_data <-
fabricate(
  cities = level(N = 2, elevation = runif(n = N, min = 1000, max = 2000)),
  citizens = level(N = 3, income = round(elevation * rnorm(n = N, mean = 5)))
)

# Specify the levels you wish to resample one of two ways:
my_data_2 <- resample_data(my_data, N = c(3, 5), ID_labels = c("cities", "citizens"))
my_data_2

my_data_3 <- resample_data(my_data, N = c(cities=3, citizens=5))
my_data_3

# Transparently pass through all units at a given level
# This example will resample 2 citizens at each of the cities:

passthrough_resample_data <- resample_data(my_data, N = c(cities=ALL, citizens=2))
passthrough_resample_data


}
