% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential_outcomes_reveal_outcomes.R
\name{reveal_outcomes}
\alias{reveal_outcomes}
\title{Reveal outcomes}
\usage{
reveal_outcomes(x)
}
\arguments{
\item{x}{A formula with the outcome name on the left hand side and the assignment variable on the right hand side (e.g., \code{Y ~ Z}). The LHS can include multiple variables (e.g., \code{c(Y1, Y2) ~ Z}).}
}
\description{
Implements a generalized switching equation. Reveals observed outcomes from multiple potential outcomes variables and an assignment variable.
}
\examples{

dat <- fabricate(
  N = 10,
  U = rnorm(N),
  potential_outcomes(Y ~ 0.1 * Z + U)
)

dat \%>\%
  fabricate(
    Z = rbinom(N, 1, prob = 0.5),
    Y = reveal_outcomes(Y ~ Z)
  )

}
