% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_categorical.R
\name{split_quantile}
\alias{split_quantile}
\title{Split data into quantile buckets}
\usage{
split_quantile(x, type)
}
\arguments{
\item{x}{Numeric vector.}

\item{type}{Number of buckets (2 = median split, 3 = terciles, etc.).}
}
\value{
Ordered factor with levels 1 through \code{type}.
}
\description{
Split data into quantile buckets
}
\examples{
split_quantile(rnorm(100), type = 4)

}
